refactor: db가 필요 없는 테스트 위치 수정
