+++ back/babble/To-Do.md
@@ -30,6 +30,6 @@
 
 ## 방 퇴장
 - [x] 방의 게스트 리스트에서 요청한 유저를 뺀다.
-- 해당 방에 요청한 유저가 존재하지 않으면 예외 처리 한다.
+- [x] 해당 방에 요청한 유저가 존재하지 않으면 예외 처리 한다.
 - 호스트가 퇴장할 경우 가장 먼저 들어온 게스트가 호스트가 된다.
 - 모든 유저가 방을 나갈 경우 방을 삭제한다.
