+++ back/babble/src/test/java/gg/babble/babble/domain/repository/RoomRepositoryTest.java
@@ -20,7 +20,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.transaction.annotation.Transactional;
 
 public class RoomRepositoryTest extends ApplicationTest {
 
@@ -35,13 +34,6 @@ public class RoomRepositoryTest extends ApplicationTest {
 
     @Autowired
     private TagService tagService;
-
-    private Room room;
-
-    @BeforeEach
-    void setUp() {
-        room = roomRepository.findById(1L).orElseThrow(BabbleNotFoundException::new);
-    }
 
     @DisplayName("방 더미 데이터를 확인한다.")
     @Test
@@ -105,60 +97,4 @@ public class RoomRepositoryTest extends ApplicationTest {
         assertThat(room.getCreatedDate()).isNotNull();
     }
 
-    @DisplayName("방에 유저가 입장한다.")
-    @Test
-    void joinRoom() {
-        User guest = User.builder()
-            .id(2L)
-            .name("손님")
-            .build();
-
-        room.join(guest);
-
-        assertThat(room.getGuests()).hasSize(1);
-        assertThat(room.getGuests()).contains(guest);
-        assertThat(guest.getRoom()).isEqualTo(room);
-    }
-
-    @DisplayName("유저가 이미 방에 있을 때 다시 유저가 방 입장을 요청하면 예외가 발생한다.")
-    @Test
-    void alreadyJoin() {
-        User guest = User.builder()
-                .id(2L)
-                .name("손님")
-                .build();
-
-        room.join(guest);
-
-        assertThatThrownBy(() -> room.join(guest)).isInstanceOf(BabbleDuplicatedException.class);
-    }
-
-    @DisplayName("유저가 방에서 나간다.")
-    @Test
-    void leave() {
-        User guest1 = User.builder()
-                .id(2L)
-                .name("손님")
-                .build();
-        User guest2 = User.builder()
-                .id(3L)
-                .name("손님")
-                .build();
-
-        room.join(guest1);
-        room.join(guest2);
-        room.leave(guest1);
-
-        assertThat(room.getGuests()).hasSize(1);
-        assertThat(guest1.getRoom()).isNull();
-    }
-
-    @DisplayName("존재하지 않는 유저가 방을 나갈 때 예외 처리한다.")
-    @Test
-    void leavingUserNotFoundInRoom() {
-        User user = userService.findById(2L);
-
-        assertThatThrownBy(() -> room.leave(user))
-            .isInstanceOf(BabbleNotFoundException.class);
-    }
 }
